requirements.txt
langchain>=0.2.17
openai>=1.44.0
streamlit>=1.38.0
python-dotenv>=1.0.1
chromadb>=0.5.3
jinja2>=3.1.4
tiktoken
markdown
python-docx>=1.1.0
pdfkit>=1.0.0
weasyprint>=61.2
cairo>=1.16
xhtml2pdf>=0.2.15

